% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk.inference.plot}
\alias{mk.inference.plot}
\title{Plot the output of a fitted hypermk model (one mk.inference experiment)}
\usage{
mk.inference.plot(fitted.obj, flux.threshold = 0, full.title = FALSE)
}
\arguments{
\item{fitted.obj}{(hypermk model) the fitted model}

\item{flux.threshold}{(double, default 0) threshold to prune low fluxes}

\item{full.title}{(Boolean, default FALSE) give full title information}
}
\value{
a ggraph plot
}
\description{
Plot the output of a fitted hypermk model (one mk.inference experiment)
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = mk_infer_cross_sectional(m)
mk.inference.plot(fit, full.title=TRUE)

}
