% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk.inference}
\alias{mk.inference}
\title{General wrapper function for HyperMk inference}
\usage{
mk.inference(
  mk.tree,
  L,
  use.priors,
  tips,
  reversible,
  optim_max_iterations = 200,
  Ntrials = 1,
  optim_algorithm = c("optim", "nlminb"),
  Nthreads = 1,
  to.nullify = matrix(nrow = 0, ncol = 2),
  flux.samples = 10000
)
}
\arguments{
\item{mk.tree}{(tree) a phylogenetic tree (which can be used to reflect independence)}

\item{L}{(integer) number of features to consider}

\item{use.priors}{(Boolean) whether we are passing prior distributions on tips or specific observations}

\item{tips}{(matrix) XXX}

\item{reversible}{(Boolean) reversible model or irreversible?}

\item{optim_max_iterations}{(integer, default 200) number of iterations before termination}

\item{Ntrials}{(integer, default 1) number of repeated trials for the inference process}

\item{optim_algorithm}{("optim" or "nlminb", default "optim") which algorithm to use}

\item{Nthreads}{(integer, default 1) number of threads to use}

\item{to.nullify}{(n x 2 matrix) from-to transitions to set to zero}

\item{flux.samples}{(integer, default 10000) number of samples to take in characterising flux}
}
\value{
a fitted hypermk model
}
\description{
takes tree, number of features, whether to use priors (vs specific states),
tip labels (priors or states), and whether to run reversible model or not
}
\details{
because setting up the data is a bit awkward, this is mainly meant for lower-level functionality

returns the fitted model with summary dataframes of transitions and fluxes
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
data = mk_cross_sectional(m, 3)
fit = mk.inference(data$tree, 3, use.priors=TRUE, data$tips, reversible = TRUE)

}
