% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_cross_sectional}
\alias{mk_cross_sectional}
\title{Cast cross-sectional data into appropriate format for Mk model}
\usage{
mk_cross_sectional(state.list, L, decimal.labels = FALSE)
}
\arguments{
\item{state.list}{(matrix) observation data. Can be binary or decimal (see below)}

\item{L}{(integer) number of features to consider}

\item{decimal.labels}{(Boolean, default FALSE) if FALSE, interpret state.list as a set of rows,
each containing a binary description of the state. if TRUE, interpret state.list as 1-indexed
decimal representations of states.}
}
\value{
a named list containing a "tree" structure and priors on the tips (as required for Mk model)
}
\description{
Cast cross-sectional data into appropriate format for Mk model
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
dataset = mk_cross_sectional(m, 3)

}
