% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_index_matrix}
\alias{mk_index_matrix}
\title{Make an indexing matrix containing required parameter structure}
\usage{
mk_index_matrix(L, reversible = TRUE, to.nullify = matrix(nrow = 0, ncol = 2))
}
\arguments{
\item{L}{(integer) number of features}

\item{reversible}{(Boolean, default TRUE) whether to create a reversible model or not}

\item{to.nullify}{(n x 2 matrix) from-to transitions to set to zero}
}
\value{
a matrix with the required parameter index structure
}
\description{
Construct matrix assigning indices to elements describing possible transitions
}
\examples{
mk_index_matrix(3)

}
