% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_infer_cross_sectional}
\alias{mk_infer_cross_sectional}
\title{Simple wrapper to immediately do inference on a matrix of cross-sectional observations}
\usage{
mk_infer_cross_sectional(m, reversible = TRUE, ...)
}
\arguments{
\item{m}{(matrix) matrix of binary observations}

\item{reversible}{(Boolean, default TRUE) reversible model?}

\item{...}{other parameters to mk.inference}
}
\value{
a fitted hypermk model
}
\description{
Simple wrapper to immediately do inference on a matrix of cross-sectional observations
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = mk_infer_cross_sectional(m)

}
