% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_infer_phylogenetic}
\alias{mk_infer_phylogenetic}
\title{Simple wrapper to immediately do inference on a matrix of observations linked by a given phylogeny}
\usage{
mk_infer_phylogenetic(m, tree, reversible = TRUE, ...)
}
\arguments{
\item{m}{(matrix) matrix of binary observations}

\item{tree}{(tree) phylogeny linking observations}

\item{reversible}{(Boolean, default TRUE) reversible model?}

\item{...}{other parameters to mk.inference}
}
\value{
a fitted hypermk model
}
\description{
Simple wrapper to immediately do inference on a matrix of observations linked by a given phylogeny
}
\examples{

m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
tree = ape::rphylo(3, 1, 1)
fit = mk_infer_phylogenetic(m, tree)

}
