% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_phylogeny_precise}
\alias{mk_phylogeny_precise}
\title{Cast phylogeny and set of tip labels data into appropriate format for Mk model}
\usage{
mk_phylogeny_precise(state.list, tree, L, decimal.labels = FALSE)
}
\arguments{
\item{state.list}{(matrix) observation data. Can be binary or decimal (see below)}

\item{tree}{(tree) describing the phylogenetic relationship of observations}

\item{L}{(integer) number of features to consider}

\item{decimal.labels}{(Boolean, default FALSE) if FALSE, interpret state.list as a set of rows,
each containing a binary description of the state. if TRUE, interpret state.list as 1-indexed
decimal representations of states.}
}
\value{
a named list containing a "tree" structure and priors on the tips (as required for Mk model)
}
\description{
Cast phylogeny and set of tip labels data into appropriate format for Mk model
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
tree = ape::rphylo(3, 1, 1)
dataset = mk_phylogeny_precise(m, tree, 3)
}
