% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_prune_model}
\alias{mk_prune_model}
\title{Prune a fitted hypermk model by removing zero-flux edges}
\usage{
mk_prune_model(fitted.obj, flux.threshold = 0)
}
\arguments{
\item{fitted.obj}{(hypermk model) the fitted model}

\item{flux.threshold}{(double, default 0) threshold to prune low fluxes}
}
\value{
a fitted, pruned model
}
\description{
Prune a fitted hypermk model by removing zero-flux edges
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
tree = ape::rphylo(3, 1, 1)
fit = mk_infer_phylogenetic(m, tree)
fit.pruned = mk_prune_model(fit)

}
