% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_pull_transitions}
\alias{mk_pull_transitions}
\title{Extract transitions as data frame from fitted Mk model}
\usage{
mk_pull_transitions(fitted.obj, reversible = TRUE)
}
\arguments{
\item{fitted.obj}{(hypermk model fit) the fitted model}

\item{reversible}{(Boolean, default TRUE) whether to use a reversible model or not}
}
\value{
a dataframe containing the transitions and their probabilities
}
\description{
Extract transitions as data frame from fitted Mk model
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = mk_infer_cross_sectional(m)
mk_pull_transitions(fit)

}
