% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{mk_simulate_fluxes}
\alias{mk_simulate_fluxes}
\title{Simulate evolutionary trajectories on a fitted Mk model}
\usage{
mk_simulate_fluxes(fitted.obj, L = 0, reversible = FALSE, nwalker = 10000)
}
\arguments{
\item{fitted.obj}{(hypermk model fit) the fitted model}

\item{L}{(default read from model fit) number of features. The ability to specify this is required for a lower-level use of this function and shouldn't be necessary.}

\item{reversible}{(default read from model fit) reversible? The ability to specify this is required for a lower-level use of this function and shouldn't be necessary.}

\item{nwalker}{(integer, default 1e4) number of walkers to simulate}
}
\value{
a dataframe containing the simulated transitions
}
\description{
Simulate evolutionary trajectories on a fitted Mk model
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = mk_infer_cross_sectional(m)
mk_simulate_fluxes(fit)

}
