% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-shared.R
\name{plot_hypercube2}
\alias{plot_hypercube2}
\title{Plot a collection of transitions as a graph. Mainly meant for lower-level functionality.}
\usage{
plot_hypercube2(trans.f, L, rates = FALSE)
}
\arguments{
\item{trans.f}{(data frame) transitions, with From/To/Flux/Rate columns}

\item{L}{(integer) number of features to consider}

\item{rates}{(Boolean, default FALSE) use rates (as opposed to fluxes)?}
}
\value{
a ggraph plot
}
\description{
Plot a collection of transitions as a graph. Mainly meant for lower-level functionality.
}
\examples{
m = matrix(c(0,0,1, 0,1,1, 1,1,1), ncol=3, nrow=3)
fit = mk_infer_cross_sectional(m)
plot_hypercube2(fit$mk_fluxes, fit$L)

}
